requirements.txt
openai==1.26.0
python-dotenv==1.0.0
langchain==0.2.2
langchain-core==0.2.4
langchain-openai==0.1.8
langchain-community==0.2.4
tiktoken==0.7.0
# chainlit==1.0.506 
chainlit==1.1.306
azure-data-tables==12.5.0
azure-ai-documentintelligence==1.0.0b3
azure-search-documents==11.4.0
# python-docx==1.1.2
bayoo-docx==0.2.20
azure-identity==1.14.0
azure-storage-blob==12.17.0
# azure-ai-formrecognizer==3.3.3
Pydantic==2.8.2
fastapi==0.110.3